Ship final spec file instead of template in release tarballs
This is easier to handle downstream, as the spec file can be used as-is,
instead of having to get rebuilt.

Adjust the clean rule to only remove the .spec if the .spec.in exists,
so that it gets removed in the upstream git, but not in unpacked release
tarball trees.