Fix webpack result test in dist/
When building in production mode, there is no index.js, just an
index.min.js.gz. This previously caused unnecessary rebuilds, or even
RPM build failures in environments where npm was not available. Test for
index.html instead, which is reliable.

Also factor this out into a constant to make it easier to adjust in
forked projects.