Move eslint/stylelint to static-code tests
stylelint 16 causes a deadlock/crash in esbuild [1]. There's no movement
on that issue, and we want stylelint 16 to be able to support the
Browser Baseline 2023.

Follow https://github.com/cockpit-project/cockpit/commit/1a28b31828e and
move the linters into test/static-code. This also starts to lint files
which are outside of esbuild's realm, such as documentation and
examples.

[1] https://github.com/evanw/esbuild/issues/3542